+++ menus/programs/beta-old.sh
@@ -1,42 +1,50 @@
-#!/bin/bash
+ #!/bin/bash
 export NCURSES_NO_UTF8_ACS=1
-clear
 
-while [ 1 ]
-do
-CHOICE=$(
-whiptail --title "Beta Menu" --menu "Make your choice" 14 60 7 \
-    "1)" "VPN Torrent - New way"   \
-    "2)" "VPN Torrent - Old way"   \
-    "3)" "DO NOT USE - For Developers Use Only!"   \
-    "4)" "Exit  "  3>&2 2>&1 1>&3
-)
+ HEIGHT=11
+ WIDTH=55
+ CHOICE_HEIGHT=5
+ BACKTITLE="Visit PlexGuide.com - Automations Made Simple"
+ TITLE="Applications - VPN Programs"
 
-result=$(whoami)
+ OPTIONS=(A "VPN Torrent"
+          B "DO NOT USE - For Developers Use Only!"
+          C "Duplicati - Advanced Backup"
+          D "PLEXTEST"
+          Z "Exit")
+
+ CHOICE=$(dialog --clear \
+                 --backtitle "$BACKTITLE" \
+                 --title "$TITLE" \
+                 --menu "$MENU" \
+                 $HEIGHT $WIDTH $CHOICE_HEIGHT \
+                 "${OPTIONS[@]}" \
+                 2>&1 >/dev/tty)
+
 case $CHOICE in
+     A)
+        bash /opt/plexguide/menus/programs/vpn.sh ;;
+     B)
+     clear
+     bash /opt/plexguide/scripts/test/move.sh
+     echo "Testing files have now been swapped"
+     echo "Please go back to the main menu to see changes"
+     read -n 1 -s -r -p "Press any key to continue "
+     ;;
+     C)
+         display=Duplicati
+         dialog --infobox "Installing: $display" 3 30
+         ansible-playbook /opt/plexguide/ansible/plexguide.yml --tags duplicati &>/dev/null &
+         sleep 2
+         dialog --msgbox 'Duplicati access: domain.com:8200 Remember to set password' 8 30
+         cronskip="yes"
+         ;;
 
-    "1)")
-    clear
-    bash /opt/plexguide/menus/programs/vpn-next.sh
-    ;;
-
-   "2)")
-   clear
-   bash /opt/plexguide/scripts/torrentvpn/torrentvpn-menu.sh
-   ;;
-
-   "3)")
-   clear
-   bash /opt/plexguide/scripts/test/move.sh
-   echo "Testing files have now been swapped"
-   echo "Please go back to the main menu to see changes"
-   read -n 1 -s -r -p "Press any key to continue "
-   ;;
-
-   "4)")
-    clear
-    exit 0
-    ;;
+     D)
+     bash /opt/plexguide/menus/plex/test.sh ;;
+     Z)
+        clear
+        exit 0 ;;
 esac
-done
-exit
+
+bash /opt/plexguide/menus/programs/beta.sh