+++ menus/programs/beta.sh
@@ -7,10 +7,9 @@ export NCURSES_NO_UTF8_ACS=1
  BACKTITLE="Visit PlexGuide.com - Automations Made Simple"
  TITLE="Applications - VPN Programs"
 
- OPTIONS=(A "VPN Torrent"
-          B "DO NOT USE - For Developers Use Only!"
-          C "Duplicati - Advanced Backup"
-          D "PLEXTEST"
+ OPTIONS=(A "DO NOT USE - For Developers Use Only!"
+          B "Duplicati - Advanced Backup"
+          C "PLEXTEST"
           Z "Exit")
 
  CHOICE=$(dialog --clear \
@@ -23,15 +22,13 @@ export NCURSES_NO_UTF8_ACS=1
 
 case $CHOICE in
      A)
-        bash /opt/plexguide/menus/programs/vpn.sh ;;
-     B)
      clear
      bash /opt/plexguide/scripts/test/move.sh
      echo "Testing files have now been swapped"
      echo "Please go back to the main menu to see changes"
      read -n 1 -s -r -p "Press any key to continue "
      ;;
-     C)
+     B)
          display=Duplicati
          dialog --infobox "Installing: $display" 3 30
          ansible-playbook /opt/plexguide/ansible/plexguide.yml --tags duplicati &>/dev/null &
@@ -40,7 +37,7 @@ case $CHOICE in
          cronskip="yes"
          ;;
 
-     D)
+     C)
      bash /opt/plexguide/menus/plex/test.sh ;;
      Z)
         clear